% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers}
\alias{markers}
\title{MGI markers}
\usage{
markers(
  query,
  var = c("mrk_id", "mrk_symbol", "entrez_gene_id", "ensembl_gene_id")
)
}
\arguments{
\item{query}{A character vector of query terms to be used in the filtering
of the returned data set of MGI markers.}

\item{var}{A string, one of: \code{"mrk_id"}, for marker identifiers;
\code{"mrk_symbol"}, for marker symbols; \code{"entrez_gene_id"}, for Entrez gene
identifiers, or \code{"ensembl_gene_id"} for Ensembl gene identifiers.}
}
\value{
A \link[tibble:tibble-package]{tibble} of 15 variables:
\itemize{
\item \code{mrk_id}: MGI marker identifier;
\item \code{mrk_type}: MGI marker type;
\item \code{mrk_symbol}: MGI marker symbol;
\item \code{mrk_name}: MGI marker name.
\item \code{genome_build}: Mouse genome build: always \code{"GRCm39"}.
\item \code{entrez_gene_id}: Entrez gene identifier.
\item \code{ncbi_gene_chr}: Chromosome name, following NCBI's gene model.
\item \code{ncbi_gene_start}: Chromosome start position, following NCBI's gene model.
\item \code{ncbi_gene_end}: Chromosome end position, following NCBI's gene model.
\item \code{ncbi_gene_strand}: Chromosome strand, following NCBI's gene model.
\item \code{ensembl_gene_id}: Ensembl gene identifier.
\item \code{ensembl_gene_chr}: Chromosome name, following Ensembl's gene model.
\item \code{ensembl_gene_start}: Chromosome start position, following Ensembl's gene model.
\item \code{ensembl_gene_end}: Chromosome end position, following Ensembl's gene model.
\item \code{ensembl_gene_strand}: Chromosome strand, following Ensembl's gene model.
}
}
\description{
\code{\link[=markers]{markers()}} provides a data set of MGI markers that have a current
associated genomic location. By current we mean as of 27 May 2024.

Use the \code{query} parameter to retrieve only a subset of the data. The query
can be any of:
\itemize{
\item marker identifier, if \code{var} is \code{"mrk_id"} (default);
\item marker symbol, if \code{var} is \code{"mrk_symbol"};
\item Entrez gene identifier, if \code{var} is \code{"entrez_gene_id"};
\item Ensembl gene identifier, if \code{var} is \code{"ensembl_gene_id"}.
}
}
\examples{
# All markers.
markers()

# actin, beta
markers("Actb", "mrk_symbol")

}
