% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi.20240527-package.R
\docType{package}
\name{mgi.20240527-package}
\alias{mgi.20240527}
\alias{mgi.20240527-package}
\title{mgi.20240527: MGI Marker Symbol and Identifiers}
\description{
MGI marker symbol and identifiers and easy crosswalks to Ensembl and Entrez gene identifiers.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
