% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi.20240527-package.R
\docType{package}
\name{mgi.20240527-package}
\alias{mgi.20240527}
\alias{mgi.20240527-package}
\title{mgi.20240527: MGI Marker Symbol and Identifiers}
\description{
Mouse Genome Informatics (MGI) marker symbol and identifiers, and easy crosswalks to Ensembl and Entrez gene identifiers based on the reports: list of mouse genetic markers (MRK_List2.rpt) and MGI gene model coordinates (MGI_Gene_Model_Coord.rpt), obtained from the MGI database by Baldarelli et al. (2024) \url{https://www.informatics.jax.org/downloads/reports}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.pattern.institute/mgi.20240527/}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
