% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_map.R
\name{symbol_map}
\alias{symbol_map}
\title{MGI marker mapping}
\usage{
symbol_map()
}
\value{
A \link[tibble:tibble-package]{tibble} with four variables:
\itemize{
\item \code{from}: A "from" marker symbol.
\item \code{to}: A "to" marker symbol, i.e. the most recent marker symbol. In most
cases this coincides with the symbol in \code{from} but not always.
\item \code{mrk_id}: Marker identifier.
\item \code{mrk_name}: Marker name.
}
}
\description{
\code{\link[=symbol_map]{symbol_map()}} provides a map of MGI marker symbols into the most recent
symbol, which can be the same, and to MGI marker identifiers.
}
\examples{
symbol_map()

}
