% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{update_symbol}
\alias{update_symbol}
\alias{symbol_to_marker_id}
\alias{marker_id_to_ensembl_id}
\alias{ensembl_id_to_marker_id}
\alias{marker_id_to_entrez_id}
\alias{entrez_id_to_marker_id}
\title{Crosswalks involving gene markers}
\usage{
update_symbol(symbol)

symbol_to_marker_id(symbol)

marker_id_to_ensembl_id(marker_id)

ensembl_id_to_marker_id(ensembl_id)

marker_id_to_entrez_id(marker_id)

entrez_id_to_marker_id(entrez_id)
}
\arguments{
\item{symbol}{A character vector of marker symbols.}

\item{marker_id}{A character vector of MGI marker identifiers.}

\item{ensembl_id}{A character vector of Ensembl gene identifiers.}

\item{entrez_id}{A character vector of Entrez gene identifiers.}
}
\value{
A character of most recent marker symbols or identifiers.
}
\description{
\itemize{
\item \code{\link[=update_symbol]{update_symbol()}} maps (potentially old) symbols to most up-to-date symbols.
\item \code{\link[=symbol_to_marker_id]{symbol_to_marker_id()}} maps (potentially old) symbols to marker identifiers.
\item \code{\link[=marker_id_to_ensembl_id]{marker_id_to_ensembl_id()}} maps MGI gene markers identifiers to an Ensembl gene identifiers.
\item \code{\link[=ensembl_id_to_marker_id]{ensembl_id_to_marker_id()}} maps Ensembl gene identifiers to MGI gene markers identifiers.
\item \code{\link[=marker_id_to_entrez_id]{marker_id_to_entrez_id()}} maps MGI gene markers identifiers to an Entrez gene identifiers.
\item \code{\link[=entrez_id_to_marker_id]{entrez_id_to_marker_id()}} maps Entrez gene identifiers to MGI gene markers identifiers.
}
}
\examples{
symbols <- c("Xkr4", "LOC102640625", "LOC102640548", "Rp1", "Sox17",
             "Gm7357", "Gm7369", "LOC102640704", "LOC102631571", "Mrpl15")

update_symbol(symbols)

symbol_to_marker_id(symbols)

}
